+++ src/pyglet2d.py
@@ -252,6 +252,23 @@ class Shape:
         """
         self.poly.shift(*vector)
 
+    def rotate(self, angle, center=None):
+        """Rotate the shape, in-place.
+
+        Parameters
+        ----------
+        angle : float
+            Angle to rotate, in radians counter-clockwise.
+        center : array-like, optional
+            Point about which to rotate.
+            If not passed, the center of the shape will be used.
+
+        """
+        args = [angle]
+        if center is not None:
+            args.extend(center)
+        self.poly.rotate(*args)
+
     def _get_vertex_list(self):
         indices = []
         for i in range(1, len(self) + 1):